refactor: extract runtime versioning helpers
Move version metadata discovery, validation, serialization, alias-path
handling, and field-name normalization into a cycle-safe leaf module.
Keep runtime orchestration and public call paths unchanged.

Validation:
- uv run make ci: passed; 856 passed; 95.83% coverage
- Python 3.12 / Pydantic 2.12.3: 856 passed; 95.73% coverage
- strict docs and immutable v0.3.0 contract: passed

Closes #108

Refs #94
